backend/requirements/constraints.txt
# Python 3.11 compatibility constraints
# These versions are pinned to ensure compatibility with Python 3.11

# Gevent and dependencies
gevent==23.9.1
greenlet==3.0.1

# Socket.IO ecosystem
Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-engineio==4.8.0
simple-websocket==1.0.0

# Email validation
email-validator==2.1.0

# Additional pinned packages to ensure stability
Werkzeug==2.3.8
SQLAlchemy==2.0.43
alembic==1.12.0
